Refactor Pedido creation to use UsuarioDTO
Updated CreatePedidoDTO to use UsuarioDTO instead of id_usuario and changed UsuarioDTO's id_usuario type from Integer to Long. Modified PedidoService to fetch Usuario entity using UsuarioDTO for Pedido creation, improving type safety and DTO consistency.

+++ src/main/java/org/example/topdeckapi/src/DTOs/DTO/UsuarioDTO.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class UsuarioDTO {
-    private Integer id_usuario;
+    private Long id_usuario;
     private String nombre;
     private String email;
     private String telefono;
